Refactor: move load album artwork data to music class when init
